Split client tests
